feat: Support for Apache Kafka protocols
This commit is an initial rough outline of how this generator
template could be extended to support another protocol.

I've tried to structure this in a way that is consistent with my
understanding of the current approach, however I'm open to
feedback about anything I've misunderstood!

I've moved all of the files that contained things that Kafka does
differently to MQ into a folder, with an index.js in-front that
loads the protocol-specific implementation. I've copied the
approach for identifying the protocol from an existing function
I found in utils. This approach does mean the protocol checking
is done multiple times, which is a little redundant. But it did
let me make the changes as transparent as possible and minimize
the changes to the existing app structure. I've also tried to do
it in a way that will make adding future protocols easy - adding
mqtt support should be trivial after this.

If this approach is acceptable, I can work on adding tests, but
I wanted to check you were all happy with the basic idea first.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>