fix: support for AsyncAPI docs with only publish or subscribe ops
The Demo class depends on channel.value.publish being null to
identify a channel without a publish operation. This assumption
doesn't work, because even for channels without a publish op
channel.value.publish is a function that returns null - it isn't
null itself. This results in uncaught exceptions when generating
code for AsyncAPI docs with only a publish or subscribe operation

This commit corrects that by replacing the checks with the
hasPublish() and hasSubscribe() methods which is a safer way to
identify if the operations exist.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>