fix: correct conversion of AsyncAPI types into Java types
I think the existing implementation was confusing "format" and
"type", comparing the type property of AsyncAPI data objects with
values that are supposed to be formats (e.g. "password").

I've reimplemented this based on the table in
https://www.asyncapi.com/docs/reference/specification/v2.5.0#dataTypeFormat

I'm using the optional format specifier where possible, as it's
more specific - and if we can't find anything matching using that
I'm falling back to the more general type specifier.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>