feat: generate schema from example message
The generator has been implemented assuming that all message objects
include a payload with properties. This is used when generating Java
classes to be able to create attributes that matching the expected
message properties.

This causes the generator to fail with AsyncAPI documents that don't
include payload properties. Some such documents will provide a
sample message instead. For these documents, we can attempt to
derive a schema from the sample, and use that schema to generate a
usable set of properties.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>